requirements.txt
Django==5.2.6
environ==1.0
django-environ==0.4.5
psycopg2-binary==2.9.9
djangorestframework==3.16.1
redis==5.2.0
channels==4.2.0
channels-redis==4.2.1
daphne==4.2.1
asgiref==3.8.1
requests==2.32.3
celery==5.5.3
django-celery-beat=2.8.1
flower==0.9.3
beautifulsoup4==4.13.5